refactor: restructure domain entities for better architecture
- Move DiagnosticProgressEntity to diagnostic_flow.dart with proper imports
- Add comprehensive DiagnosticStage enum with Portuguese display names
- Remove DiagnosticProgressEntity duplicate from final_results_entity.dart
- Create DiagnosticProgressModel for JSON serialization needs
- Add speedTestResult optional field to DiagnosticProgressEntity
- Remove redundant file header comments for cleaner code
- Improve entity separation and reduce circular dependencies
- Support for progress tracking with final results integration
